+++ django_for_agents/templatetags/for_agents.py
@@ -5,6 +5,7 @@ from typing import Any
 from django import template
 from django.http import HttpRequest
 
+
 register = template.Library()